requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
requests
numpy
pandas
scikit-learn
tensorflow
torch
torchvision
joblib
supabase
gunicorn
whitenoise
